Updated syntax
Updated the syntax for a lot of the more complex values such as Long REQ, Med REQ, and Request and/or Answer. Provided a method to handle how to print most of them. Also, finally reduced the handle_first_page() if-else statements in a way that generates the same outcome.