Updated Syntax Major Change for Values
Simple updates to some syntax related questions. Major changes for values by changing the way information is passed to the methods and logically separated in the report structure. Was able to shorten the code by a little bit and make a cleaner way to update in the future. Added three new if statements to the third page generation for logically different questions than previously used too.